windows script fix

+++ Windows/prepare.ps1
@@ -85,14 +85,18 @@ function Font-Setup {
     "${env:LOCALAPPDATA}\Microsoft\Windows\Fonts" *> ${null}
 
   Get-ChildItem -Path "${HOME}\setup\Assets\${FONT}" -Filter "*.ttf" | ForEach-Object {
-    Copy-Item -Force `
-      -Path $($_.FullName) `
-      -Destination "${env:LOCALAPPDATA}\Microsoft\Windows\Fonts\$($_.Name)"
+    if (-not (Test-Path "${env:LOCALAPPDATA}\Microsoft\Windows\Fonts\$($_.Name)")) {
+      Copy-Item -Force `
+        -Path $($_.FullName) `
+        -Destination "${env:LOCALAPPDATA}\Microsoft\Windows\Fonts\$($_.Name)"
+    }
 
-    Set-ItemProperty -Force `
-      -Path "HKCU:\SOFTWARE\Microsoft\Windows NT\CurrentVersion\Fonts" `
-      -Name "$([System.IO.Path]::GetFileNameWithoutExtension($_.Name) -replace '-', ' ') (TrueType)" `
-      -Value "${env:LOCALAPPDATA}\Microsoft\Windows\Fonts\$($_.Name)"
+    if (Test-Path "${env:LOCALAPPDATA}\Microsoft\Windows\Fonts\$($_.Name)") {
+      Set-ItemProperty -Force `
+        -Path "HKCU:\SOFTWARE\Microsoft\Windows NT\CurrentVersion\Fonts" `
+        -Name "$([System.IO.Path]::GetFileNameWithoutExtension($_.Name) -replace '-', ' ') (TrueType)" `
+        -Value "${env:LOCALAPPDATA}\Microsoft\Windows\Fonts\$($_.Name)"
+    }
   }
 }
